feat: use container-init.d mechanism to initialize config
The upstream image from Kubo now has a mechanism to run an init script
before the daemon starts. Instead of replacing the start_ipfs script
with our own we now use the upstream script and then add in our own init
script. This enable further downstream consumers of this image to
further customize the configuration.