feat: use container-init.d mechanism to initialize config (#24)
* feat: use container-init.d mechanism to initialize config

The upstream image from Kubo now has a mechanism to run an init script
before the daemon starts. Instead of replacing the start_ipfs script
with our own we now use the upstream script and then add in our own init
script. This enable further downstream consumers of this image to
further customize the configuration.

* fix: use ipfs-config-identity locally

Prior to this change we downloaded this script from a github gist.
This adds an unnecessary point of failure and security risk. Now the
contents are simply part of the repo.

* fix: remove gateway logic

With this change the gateway is disabled instead of exposed.